requirements.txt
asgiref==3.5.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.0.2
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
flake8-html==0.4.1
idna==3.4
importlib-metadata==4.10.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.7
urllib3==1.26.13
zipp==3.7.0
